Updated materials and names for compatability
Kidney was named KidneyOne, which wouldn't match the appropriate enum in code. Also changed the rest of the materials to start as Opaque, since transparant variants will be generated of these. I did a small update to the .gitignore to include the extra folder so it doesn't have to be manually created, the materials do still get gitignored.