Revamped ManagedMaterials Dictionary accessing
It's now based on material name instead of enum name, which is far more dynamic and easier to use. Also update a lot of the error handling to be more concise and precise, also following DRY principles. Error handling gets done during editor time and runtime. There was also a bug with the idleOrgans list being incorrectly populated due to an order of execution issue, this has also been fixed.